+++ experimental/CHANGELOG.md
@@ -12,6 +12,8 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 
 ### :rocket: Features
 
+* feat(exporter-prometheus): support configuring the default aggregation [#6761](https://github.com/open-telemetry/opentelemetry-js/pull/6761) @ArthurSens
+
 ### :bug: Bug Fixes
 
 ### :books: Documentation